Allow a "friendly hostname" to be configured
If this is set, the module is also responsible for arranging for the
configured API to accept the given custom hostname, thus allowing the
registry to work mostly "out of the box" using the discovery document
served via the API.

This is optional because a user may instead elect to expose a registry
at the API gateway default hostname and then manually place a service
discovery document on a separate web server running at the desired
friendly hostname. In this case, they will use the "services" output
to get the content of that discovery document.